flask_app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.1
WTForms==3.1.1
SQLAlchemy==2.0.23
email-validator==2.1.0.post1
python-dotenv==1.0.0
flask-mail==0.10.0
flask-login-0.6.3
flask-migrate==4.1.0
email-validator-2.2.0